Move enums containing the nodeId for the machine
