+++ BrewMesModulesParent/common/pom.xml
@@ -28,6 +28,11 @@
 			<version>3.1.8</version>
 			<scope>compile</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.eclipse.milo</groupId>
+			<artifactId>stack-core</artifactId>
+			<version>0.5.2</version>
+		</dependency>
 	</dependencies>
 
 </project>
