+++ BrewMesModulesParent/common/src/main/java/com/brewmes/common/util/machinenodes/Constants.java
@@ -1,4 +1,4 @@
-package com.brewmes.subscriber.util;
+package com.brewmes.common.util.machinenodes;
 
 class Constants {
     public static final int NAMESPACE_INDEX_VALUE = 6;
