+++ BrewMesModulesParent/common/src/main/java/com/brewmes/common/util/machinenodes/StatusNodes.java
@@ -1,4 +1,4 @@
-package com.brewmes.subscriber.util;
+package com.brewmes.common.util.machinenodes;
 
 import org.eclipse.milo.opcua.stack.core.types.builtin.NodeId;
