+++ BrewMesModulesParent/common/src/test/java/com/brewmes/common/util/machinenodes/AdminNodesTest.java
@@ -1,4 +1,4 @@
-package com.brewmes.subscriber.util;
+package com.brewmes.common.util.machinenodes;
 
 import org.junit.jupiter.api.Test;
